update Postrender component
