+++ src/components/PostRenderer/index.tsx
@@ -1,10 +1,14 @@
-import { Alert, AlertTitle, Box, Divider, Typography } from "@mui/material";
+import { Divider, Typography } from "@mui/material";
 import React from "react";
+import ReactMarkdown from "react-markdown";
 import { SpecialComponents } from "react-markdown/lib/ast-to-react";
 import { NormalComponents } from "react-markdown/lib/complex-types";
+import remarkGfm from "remark-gfm";
+import remarkUnwrapImages from "remark-unwrap-images";
 import CodeBlock from "../CodeBlock";
+import Image from "next/image";
 
-const PostRenderer: Partial<
+const Components: Partial<
   Omit<NormalComponents, keyof SpecialComponents> & SpecialComponents
 > = {
   h1: ({ children }) => (
@@ -92,4 +96,37 @@ const PostRenderer: Partial<
   code: CodeBlock,
 };
 
+type Props = {
+  post: Post;
+};
+
+const PostRenderer: React.VFC<Props> = ({ post }) => {
+  const ImgRenderer: React.VFC<{ alt?: string; src?: string }> = (props) => {
+    const { alt, src } = props;
+
+    if (!src) return <></>;
+
+    const imgSrc = isURL(src) ? src : require(`posts/${post.slug}/${src}`);
+    return <Image src={imgSrc} alt={alt} />;
+  };
+
+  return (
+    <ReactMarkdown
+      components={{ img: ImgRenderer, ...Components }}
+      remarkPlugins={[remarkUnwrapImages, remarkGfm]}
+    >
+      {post.content}
+    </ReactMarkdown>
+  );
+};
+
 export default PostRenderer;
+
+const isURL = (url: string) => {
+  try {
+    new URL(url);
+  } catch (e) {
+    return false;
+  }
+  return true;
+};
